Implement post-ranking numeric candidate variants

+++ app/src/androidTest/java/com/kazumaproject/markdownhelperkeyboard/converter/candidate/ExactInputCandidatePromoterPerformanceInstrumentedTest.kt
@@ -208,7 +208,7 @@ class ExactInputCandidatePromoterPerformanceInstrumentedTest {
     ): Candidate = Candidate(
         string = string,
         type = type,
-        length = string.length.toUByte(),
+        length = string.length,
         score = 0,
     )
