+++ app/src/main/java/com/kazumaproject/markdownhelperkeyboard/converter/candidate/ZenzCandidate.kt
@@ -3,7 +3,7 @@ package com.kazumaproject.markdownhelperkeyboard.converter.candidate
 data class ZenzCandidate(
     val string: String,
     val type: Byte,
-    val length: UByte,
+    val length: Int,
     val score: Int,
     val originalString: String,
     val leftId: Short? = null,
