+++ app/src/main/java/com/kazumaproject/markdownhelperkeyboard/setting_activity/ui/key_candidate_letter_size/KeyCandidateLetterSizeFragment.kt
@@ -231,16 +231,16 @@ class KeyCandidateLetterSizeFragment : Fragment() {
             Candidate(
                 string = "プレビュー",
                 type = 1.toByte(),
-                length = (5).toUByte(),
+                length = (5),
                 score = 4000
             ),
             Candidate(
                 string = "文字サイズ",
                 type = 1.toByte(),
-                length = (5).toUByte(),
+                length = (5),
                 score = 4001
             ),
-            Candidate(string = "候補", type = 1.toByte(), length = (2).toUByte(), score = 4002)
+            Candidate(string = "候補", type = 1.toByte(), length = (2), score = 4002)
         )
         suggestionAdapter.suggestions = previewCandidates
     }
