+++ app/src/test/java/com/kazumaproject/markdownhelperkeyboard/candidate_order/CandidateOrderOverrideSavedOrderEditTest.kt
@@ -195,7 +195,7 @@ class CandidateOrderOverrideSavedOrderEditTest {
         com.kazumaproject.markdownhelperkeyboard.converter.candidate.Candidate(
             string = string,
             type = 1.toByte(),
-            length = string.length.toUByte(),
+            length = string.length,
             score = 0
         )
