+++ app/src/test/java/com/kazumaproject/markdownhelperkeyboard/ime_service/adapters/SuggestionAdapterUserTemplateTypeTest.kt
@@ -207,7 +207,7 @@ class SuggestionAdapterUserTemplateTypeTest {
         return Candidate(
             string = string,
             type = type,
-            length = string.length.toUByte(),
+            length = string.length,
             score = 0
         )
     }
