+++ app/src/test/java/com/kazumaproject/markdownhelperkeyboard/ime_service/candidate/CandidateStripClipboardPreviewRegressionTest.kt
@@ -235,7 +235,7 @@ class CandidateStripClipboardPreviewRegressionTest {
         Candidate(
             string = text,
             type = 1.toByte(),
-            length = text.length.toUByte(),
+            length = text.length,
             score = 0,
             yomi = text
         )
