+++ app/src/test/java/com/kazumaproject/markdownhelperkeyboard/ime_service/candidate/CandidateStripContentResolverTest.kt
@@ -695,7 +695,7 @@ class CandidateStripContentResolverTest {
         Candidate(
             string = text,
             type = 1.toByte(),
-            length = text.length.toUByte(),
+            length = text.length,
             score = 0,
             yomi = text
         )
